rename tableNewArgument -> tableInputArgument
